Final commit, figured out problems and simplified
