+++ node/package.json
@@ -10,8 +10,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.24.0",
-    "cron": "^1.8.2",
     "express": "^4.17.1",
     "firebase-admin": "^10.0.0",
     "node-schedule": "^2.0.0"
